FixMatch Code/google_requirements.txt

absl-py
easydict
cython
numpy
tensorflow
tqdm
scipy
pillow
torch==1.8.0
torchvision==0.9.0
torchaudio
torchtext
omegaconf
hydra
hydra-core
ignite
pytorch-ignite
tensorboardX
scikit-learn
matplotlib
abel-pytorch
